Don't clobber standalone.xml on subsequent builds

+++ build/assembly/lib/assembly_tool.rb
@@ -154,8 +154,10 @@ class AssemblyTool
 
   def rename_standalone_xml
     Dir.chdir( File.join( @jboss_dir, 'standalone', 'configuration' ) ) do
-      FileUtils.mv( 'standalone.xml', 'standalone-original.xml' )
-      FileUtils.mv( 'standalone-preview.xml', 'standalone.xml' )
+      if File.exists?( 'standalone-preview.xml' )
+        FileUtils.mv( 'standalone.xml', 'standalone-original.xml' )
+        FileUtils.mv( 'standalone-preview.xml', 'standalone.xml' )
+      end
     end
   end
